Add .env example for local development use
A 12 factor application must have its sensitive data sotred in
environment variables, but it is expensive to set by hand all
those variables on the system.

The Django-environ package used to handdle the environment, has
support for .env files. This file must be inside the settings directory.

DO NOT USE .env files on production.